api-integration/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
